% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_install.R
\name{act_install}
\alias{act_install}
\title{act_install}
\usage{
act_install(system = c("linux", "macos", "windows"))
}
\arguments{
\item{system}{chr}
}
\description{
install act and run github action locally
}
