% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesisdown.R
\docType{package}
\name{brockdown}
\alias{brockdown}
\alias{brockdown-package}
\title{brockr: A package for creating undergraduate, Masters, and PhD theses
using R Markdown}
\description{
This package creates theses/dissertations for Brock University.
}
\section{thesis_gitbook}{

Creates an R Markdown thesis template as a webpage
}

\section{thesis_pdf}{

Creates an R Markdown thesis template as a PDF
}

\section{thesis_word}{

Creates an R Markdown thesis template as a Microsoft Word document
}

