% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_install.R
\name{docker_install}
\alias{docker_install}
\title{docker_install}
\usage{
docker_install(system)
}
\arguments{
\item{system}{linux,macos,windows}
}
\value{
null
}
\description{
install docker in linux systems
}
