% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kubectl_apply.R
\name{kubectl_apply}
\alias{kubectl_apply}
\title{kubectl_apply}
\usage{
kubectl_apply(yamlfiles, namespace = NULL)
}
\arguments{
\item{yamlfiles}{chr}

\item{namespace}{chr,default as NULL}
}
\value{
yamlfiles
}
\description{
kubectl_apply functions
}
