% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kubectl_delete.R
\name{kubectl_delete}
\alias{kubectl_delete}
\title{kubectl_delete}
\usage{
kubectl_delete(object, servicetype)
}
\arguments{
\item{object}{the object to remove}

\item{servicetype}{chr,one of the following:deployments,service,ingress,namespaces,configmap,jobs, cronjob,secret,endpoints}
}
\value{
chr
}
\description{
kubectl delete warpper
}
