% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kubectl_exec.R
\name{kubectl_exec}
\alias{kubectl_exec}
\title{kubectl_exec}
\usage{
kubectl_exec(podname, arg_command)
}
\arguments{
\item{podname}{podnames}

\item{arg_command}{the command to be ran}
}
\description{
kubectl exec wrapper
}
