% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kubectl_get.R
\name{kubectl_get}
\alias{kubectl_get}
\title{kubectl_get}
\usage{
kubectl_get(servicetype)
}
\arguments{
\item{servicetype}{chr,one of the following:deployments,service,ingress,namespaces,configmap,jobs, cronjob,secret,endpoints}
}
\value{
data.frame
}
\description{
kubectl_get functions for
deployments,service,ingress,namespaces,configmap,jobs,
cronjob,secret,endpoints
}
