% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kubectl_install.R
\name{kubectl_install}
\alias{kubectl_install}
\title{kubectl_install}
\usage{
kubectl_install(system = c("linux", "macos", "windows"))
}
\arguments{
\item{system}{linux,macos,windows}
}
\value{
null
}
\description{
install kubectl in linux
}
