% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kubectl_portforward.R
\name{kubectl_portforward}
\alias{kubectl_portforward}
\title{kubectl_portforward}
\usage{
kubectl_portforward(podname, hostport, childport)
}
\arguments{
\item{podname}{the podname}

\item{hostport}{the hostport}

\item{childport}{the childport for the containers}
}
\description{
kubectl port-forward warpper
}
