% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minikube_install.R
\name{minikube_install}
\alias{minikube_install}
\title{minikube_install}
\usage{
minikube_install(system = c("linux", "macos", "windows"))
}
\arguments{
\item{system}{linux,macos,windows}
}
\value{
null
}
\description{
install minikube
}
