% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placeholder.R
\name{placeholder}
\alias{placeholder}
\title{placeholder}
\usage{
placeholder(path)
}
\arguments{
\item{path}{char,a dir to be make}
}
\value{
path
}
\description{
creat a dir that is not empty
}
\examples{
placeholder("testtemp")
fs::dir_delete("testtemp")
}
