% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_caption.R
\name{process_citations}
\alias{process_citations}
\title{Process citations in a text formatted with markdown}
\usage{
process_citations(caption, bib, csl, figname = "fig1", to = "latex")
}
\arguments{
\item{caption}{text formatted with markdown}

\item{bib}{the name of the bibliography file(s)}

\item{csl}{the name of the csl file}

\item{figname}{the name of the output file}

\item{to}{the output format. Defaults to "latex". Could also be "html".
"markdown" does nothing useful.}
}
\value{
formatted text with rendered citations
}
\description{
Process citations in a text formatted with markdown
}
\examples{
# Setup for the example
# Note that this will already be set for you when you run the document
rootdir <- find.package("beaverdown")
rootdir <- paste0(rootdir, "/rmarkdown/templates/oregonstate/skeleton/")
bib <- file.path(rootdir, c("bib/references.bib", "bib/thesis.bib"))
csl <- file.path(rootdir, "csl/apa.csl")
txt <- "**Hey!** This is a citation from @angel2000."
process_citations(txt, bib, csl)
}

