% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_job.R
\name{submit_job}
\alias{submit_job}
\title{submit_job}
\usage{
submit_job(
  task_names,
  namespace = "default",
  container,
  command,
  ncpu = 8,
  nram = "1024Mi",
  cpu_limit = 100,
  nram_limit = "10240000Mi",
  env_list = NULL,
  hostPath = getwd(),
  restartPolicy = "Never"
)
}
\arguments{
\item{task_names}{the name of the submit job}

\item{namespace}{the namespace used}

\item{container}{the container used}

\item{command}{the command submit to run}

\item{ncpu}{the cpu request}

\item{nram}{the memory request}

\item{cpu_limit}{the limit cpu}

\item{nram_limit}{the nram limit}

\item{env_list}{the env value list,default as NULL}

\item{hostPath}{the attached hostpath,default as the current file dir,attached to /mnt in the container}

\item{restartPolicy}{the restartPolicy,default as Never}
}
\description{
submit_job and run
}
